refact -> create getUpgradeUrl() util helper
