+++ extensions/shared/components/upgrade-nudge/index.jsx
@@ -3,8 +3,7 @@
  */
 import GridiconStar from 'gridicons/dist/star';
 import { __, sprintf } from '@wordpress/i18n';
-import { addQueryArgs } from '@wordpress/url';
-import { compact, get, startsWith } from 'lodash';
+import { get, startsWith } from 'lodash';
 import { compose } from '@wordpress/compose';
 import { withSelect } from '@wordpress/data';
 import { useEffect } from '@wordpress/element';
@@ -14,10 +13,9 @@ import { useEffect } from '@wordpress/element';
  */
 import analytics from '../../../../_inc/client/lib/analytics';
 import BlockNudge from '../block-nudge';
-import getSiteFragment from '../../get-site-fragment';
-import { isSimpleSite } from '../../site-type-utils';
-import './store';
+import { getUpgradeUrl } from '../../plan-utils';
 
+import './store';
 import './style.scss';
 
 const getTitle = ( customTitle, planName ) => {
@@ -94,45 +92,16 @@ export const UpgradeNudge = ( {
 export default compose( [
 	withSelect( ( select, { plan: planSlug, blockName } ) => {
 		const plan = select( 'wordpress-com/plans' ).getPlan( planSlug );
+		const postId = select( 'core/editor' ).getCurrentPostId();
+		const postType = select( 'core/editor' ).getCurrentPostType();
+
+		const upgradeUrl = getUpgradeUrl( { plan, planSlug, postId, postType } );
 
 		// WP.com plan objects have a dedicated `path_slug` field, Jetpack plan objects don't
 		// For Jetpack, we thus use the plan slug with the 'jetpack_' prefix removed.
 		const planPathSlug = startsWith( planSlug, 'jetpack_' )
 			? planSlug.substr( 'jetpack_'.length )
 			: get( plan, [ 'path_slug' ] );
-
-		const postId = select( 'core/editor' ).getCurrentPostId();
-		const postType = select( 'core/editor' ).getCurrentPostType();
-
-		// The editor for CPTs has an `edit/` route fragment prefixed
-		const postTypeEditorRoutePrefix = [ 'page', 'post' ].includes( postType ) ? '' : 'edit';
-
-		// Post-checkout: redirect back here
-		const redirect_to = isSimpleSite()
-			? addQueryArgs(
-					'/' +
-						compact( [ postTypeEditorRoutePrefix, postType, getSiteFragment(), postId ] ).join(
-							'/'
-						),
-					{
-						plan_upgraded: 1,
-					}
-			  )
-			: addQueryArgs(
-					window.location.protocol +
-						`//${ getSiteFragment().replace( '::', '/' ) }/wp-admin/post.php`,
-					{
-						action: 'edit',
-						post: postId,
-						plan_upgraded: 1,
-					}
-			  );
-
-		const upgradeUrl =
-			planPathSlug &&
-			addQueryArgs( `https://wordpress.com/checkout/${ getSiteFragment() }/${ planPathSlug }`, {
-				redirect_to,
-			} );
 
 		const planName = get( plan, [ 'product_name' ] );
 		return {
